# docs/swagger.yaml
definitions:
  db.CreateProductParams:
    properties:
      data_venda:
        type: string
      descricao:
        type: string
      nome:
        type: string
      preco:
        type: string
    type: object
  db.Product:
    properties:
      data_venda:
        type: string
      descricao:
        type: string
      id:
        type: integer
      nome:
        type: string
      preco:
        type: string
    type: object
  db.UpdateProductParams:
    properties:
      data_venda:
        type: string
      descricao:
        type: string
      id:
        type: integer
      nome:
        type: string
      preco:
        type: string
    type: object
info:
  contact: {}
paths:
  /product:
    post:
      consumes:
      - application/json
      description: Cria um produto com nome, descrição, preço e data de venda
      parameters:
      - description: Dados do produto
        in: body
        name: produto
        required: true
        schema:
          $ref: '#/definitions/db.CreateProductParams'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cria um novo produto
      tags:
      - produtos
  /product/{id}:
    delete:
      description: Exclui um produto do banco de dados
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove um produto
      tags:
      - produtos
    get:
      description: Retorna os dados de um produto específico
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Busca um produto por ID
      tags:
      - produtos
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um produto existente
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: integer
      - description: Dados atualizados
        in: body
        name: produto
        required: true
        schema:
          $ref: '#/definitions/db.UpdateProductParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualiza um produto
      tags:
      - produtos
  /products:
    get:
      description: Retorna todos os produtos cadastrados
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Product'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Lista todos os produtos
      tags:
      - produtos
swagger: "2.0"